+++ etc/docker/integration/run.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+echo "Execute for $host"
+newman run collection.json --env-var host=$host