+++ src/deps.ts
@@ -1,5 +1,1 @@
-export {
-  serve,
-  Server,
-  ServerRequest,
-} from "https://deno.land/std@0.104.0/http/server.ts";
+export { serve } from "https://deno.land/std@0.155.0/http/server.ts";